impl sumcheck compression for SumCheckProof
