+++ src/jolt_compression/mod.rs
@@ -1,3 +1,2 @@
 //! This module implements Jolt compression
 pub mod sumcheck_compression;
-pub mod sumcheck_compression2;
