+++ src/jolt_compression/sumcheck_compression.rs
@@ -1,17 +1,22 @@
-//! Compression
+//! Compression for sumcheck proofs using Groth16
 use crate::{
   frontend::{
     groth16::{create_random_proof, generate_random_parameters, Parameters, Proof},
     num::AllocatedNum, Circuit, ConstraintSystem, SynthesisError,
-  }, hypernova::ro_sumcheck::ROSumcheckProof, provider::Bn256EngineKZG, spartan::polys::univariate::UniPoly
+  },
+  spartan::{
+    polys::univariate::UniPoly,
+    sumcheck::SumcheckProof,
+  },
+  provider::Bn256EngineKZG,
 };
 use ff::PrimeField;
 use halo2curves::bn256::Fr;
 use rand::thread_rng;
 
-/// Circuit for verifying a sumcheck proof
+/// Circuit for verifying a sumcheck proof with provided challenges
 #[derive(Clone)]
-pub struct SumcheckVerificationCircuit<F: PrimeField> {
+pub struct SumcheckVerifierCircuit<F: PrimeField> {
   /// The univariate polynomials from the sumcheck proof
   pub polys: Vec<UniPoly<F>>,
   /// The claimed sum
@@ -20,22 +25,32 @@ pub struct SumcheckVerificationCircuit<F: PrimeField> {
   pub num_rounds: usize,
   /// The degree bound for the univariate polynomials
   pub degree_bound: usize,
+  /// The verifier's challenges
+  pub r: Vec<F>,
 }
 
-impl<F: PrimeField> Circuit<F> for SumcheckVerificationCircuit<F> {
+impl<F: PrimeField> Circuit<F> for SumcheckVerifierCircuit<F> {
   fn synthesize<CS: ConstraintSystem<F>>(self, cs: &mut CS) -> Result<(), SynthesisError> {
     // Allocate the claim as a public input
     let claim = AllocatedNum::alloc_input(cs.namespace(|| "claim"), || Ok(self.claim))?;
-
-    // Verify each round of the sumcheck proof
-    let mut e = claim;
 
     // Verify that there is a univariate polynomial for each round
     if self.polys.len() != self.num_rounds {
       return Err(SynthesisError::Unsatisfiable);
     }
 
-    for (i, poly) in self.polys.iter().enumerate() {
+    // Verify that there is a challenge for each round
+    if self.r.len() != self.num_rounds {
+      return Err(SynthesisError::Unsatisfiable);
+    }
+
+    // Start with the initial claim
+    let mut e = claim;
+
+    for i in 0..self.num_rounds {
+      // Get the polynomial for this round
+      let poly = &self.polys[i];
+
       // Verify degree bound
       if poly.degree() != self.degree_bound {
         return Err(SynthesisError::Unsatisfiable);
@@ -70,12 +85,9 @@ impl<F: PrimeField> Circuit<F> for SumcheckVerificationCircuit<F> {
         |lc| lc + e.get_variable(),
       );
 
-      // Derive the verifier's challenge for the next round
-      // In a real implementation, we would use a random oracle here
-      // For simplicity, we'll just use a fixed challenge
+      // Allocate the verifier's challenge for this round
       let r_i = AllocatedNum::alloc(cs.namespace(|| format!("challenge_{}", i)), || {
-        // This is a placeholder - in a real implementation, this would be derived from a transcript
-        Ok(F::from(i as u64 + 1))
+        Ok(self.r[i])
       })?;
 
       // Evaluate the polynomial at r_i
@@ -93,10 +105,12 @@ impl<F: PrimeField> Circuit<F> for SumcheckVerificationCircuit<F> {
           &term,
         )?;
 
-        power = power.mul(
-          cs.namespace(|| format!("power_update_{}_{}", i, j)),
-          &r_i,
-        )?;
+        if j < coeffs.len() - 1 {
+          power = power.mul(
+            cs.namespace(|| format!("power_update_{}_{}", i, j)),
+            &r_i,
+          )?;
+        }
       }
 
       // Update e for the next round
@@ -122,13 +136,15 @@ impl SumcheckCompression {
     claim: Fr,
     num_rounds: usize,
     degree_bound: usize,
+    r: Vec<Fr>,
   ) -> Result<(Self, Parameters<Bn256EngineKZG>), SynthesisError> {
     // Create the circuit for verifying the sumcheck proof
-    let circuit = SumcheckVerificationCircuit::<Fr> {
+    let circuit = SumcheckVerifierCircuit::<Fr> {
       polys,
       claim,
       num_rounds,
       degree_bound,
+      r,
     };
 
     // Generate parameters for the Groth16 proof system
@@ -141,17 +157,33 @@ impl SumcheckCompression {
     Ok((Self { proof }, params))
   }
 
-  /// Compress a ROSumcheckProof into a Groth16 proof
-  pub fn compress_ro_proof(
-    proof: &ROSumcheckProof<Bn256EngineKZG>,
+  /// Compress a SumcheckProof into a Groth16 proof
+  pub fn compress_proof(
+    proof: &SumcheckProof<Bn256EngineKZG>,
     claim: Fr,
     num_rounds: usize,
     degree_bound: usize,
+    r: Vec<Fr>,
   ) -> Result<(Self, Parameters<Bn256EngineKZG>), SynthesisError> {
-    // Convert ROSumcheckProof to Vec<UniPoly>
-    let polys = proof.polys.clone();
+    // Convert SumcheckProof to Vec<UniPoly> using the correct claim value for each round
+    let mut polys = Vec::with_capacity(num_rounds);
+    let mut current_claim = claim;
 
-    Self::compress(polys, claim, num_rounds, degree_bound)
+    for (i, compressed_poly) in proof.compressed_polys.iter().enumerate() {
+      // Decompress the polynomial using the current claim value
+      let poly = compressed_poly.decompress(&current_claim);
+
+      // Update the claim for the next round if not the last round
+      if i < num_rounds - 1 {
+        // Evaluate the polynomial at the challenge point to get the next claim
+        current_claim = poly.evaluate(&r[i]);
+      }
+
+      // Add the polynomial to our collection (clone it to avoid ownership issues)
+      polys.push(poly.clone());
+    }
+
+    Self::compress(polys, claim, num_rounds, degree_bound, r)
   }
 }
 
@@ -159,232 +191,76 @@ impl SumcheckCompression {
 mod tests {
   use super::*;
   use crate::{
-    frontend::groth16::{verify_proof, prepare_verifying_key},
-    spartan::polys::univariate::UniPoly
+    frontend::groth16::{prepare_verifying_key, verify_proof}, provider::Bn256EngineKZG, spartan::{
+      polys::multilinear::MultilinearPolynomial,
+      sumcheck::SumcheckProof,
+    }, traits::{Engine, TranscriptEngineTrait}
   };
   use halo2curves::bn256::Fr;
   use rand::{Rng, thread_rng};
 
   #[test]
   fn test_sumcheck_compression() {
-    // Create a simple sumcheck proof with two rounds
-    // For the sumcheck protocol to verify:
-    // 1. For the first polynomial: eval_at_zero + eval_at_one = claim (10)
-    // 2. For the second polynomial: eval_at_zero + eval_at_one = eval_poly1(r_0)
-    //    where r_0 = 1 (hardcoded in the circuit)
+    let ns = [2, 4, 8, 16];
+    let num_tries = 10;
 
-    // First polynomial: coefficients [1, 2, 6]
-    // eval_at_zero = 1
-    // eval_at_one = 1+2+6 = 9
-    // eval_at_zero + eval_at_one = 1+9 = 10 (matches claim)
-    let poly1 = UniPoly::new(vec![Fr::from(1), Fr::from(2), Fr::from(6)]);
+    for &n in &ns {
+      for t in 0..num_tries {
+        println!("Testing n={} ({}/{})...", n, t+1, num_tries);
+        // Generate random polynomials A and B
+        let mut rng = thread_rng();
+        let size = 1 << n;
 
-    // Evaluate poly1 at r_0 = 1: 1 + 2*1 + 6*1^2 = 1 + 2 + 6 = 9
-    // Second polynomial: coefficients [0, 5, 4]
-    // eval_at_zero = 0
-    // eval_at_one = 0+5+4 = 9
-    // eval_at_zero + eval_at_one = 0+9 = 9 (matches poly1(r_0))
-    let poly2 = UniPoly::new(vec![Fr::from(0), Fr::from(5), Fr::from(4)]);
+        let mut poly_a = MultilinearPolynomial::new(
+          (0..size).map(|_| Fr::from(rng.gen_range(1..10) as u64)).collect()
+        );
 
-    // Compress the proof
-    let (compressed_proof, params) = SumcheckCompression::compress(
-      vec![poly1, poly2],
-      Fr::from(10), // claim
-      2,            // num_rounds
-      2,            // degree_bound
-    ).unwrap();
+        let mut poly_b = MultilinearPolynomial::new(
+          (0..size).map(|_| Fr::from(rng.gen_range(1..10) as u64)).collect()
+        );
 
-    let pvk = prepare_verifying_key(&params.vk);
+        // Define the combination function (a * b)
+        let comb_func = |a: &Fr, b: &Fr| *a * *b;
 
-    let verified = verify_proof(&pvk, &compressed_proof.proof, &[Fr::from(10)]).unwrap();
+        // Calculate the claim (sum of all a_i * b_i)
+        let claim = (0..size)
+          .map(|i| comb_func(&poly_a[i], &poly_b[i]))
+          .fold(Fr::zero(), |acc, val| acc + val);
 
-    assert!(verified);
-  }
+        let num_rounds = n;
+        let degree_bound = 2; // For quadratic polynomials
 
+        // Create a transcript for the proof
+        let mut transcript = <Bn256EngineKZG as Engine>::TE::new(b"test");
 
-  /// Generate a valid set of polynomials for a sumcheck proof with the given claim
-  fn generate_valid_polys(num_rounds: usize, degree_bound: usize, claim: Fr) -> Vec<UniPoly<Fr>> {
-    // We'll use the exact same polynomials as in the test_sumcheck_compression test
-    // This ensures that the polynomials satisfy the constraints and will be verified successfully
+        // Generate the sumcheck proof
+        let (proof, r, _) = SumcheckProof::prove_quad(
+          &claim,
+          num_rounds,
+          &mut poly_a,
+          &mut poly_b,
+          comb_func,
+          &mut transcript,
+        ).unwrap();
 
-    if num_rounds == 2 && degree_bound == 2 {
-      // First polynomial: coefficients [1, 2, 6]
-      // eval_at_zero = 1
-      // eval_at_one = 1+2+6 = 9
-      // eval_at_zero + eval_at_one = 1+9 = 10 (matches claim)
-      let poly1 = UniPoly::new(vec![Fr::from(1), Fr::from(2), Fr::from(6)]);
+        // Verify the original proof
+        let mut verify_transcript = <Bn256EngineKZG as Engine>::TE::new(b"test");
+        assert!(proof.verify(claim, num_rounds, degree_bound, &mut verify_transcript).is_ok());
 
-      // Evaluate poly1 at r_0 = 1: 1 + 2*1 + 6*1^2 = 1 + 2 + 6 = 9
-      // Second polynomial: coefficients [0, 5, 4]
-      // eval_at_zero = 0
-      // eval_at_one = 0+5+4 = 9
-      // eval_at_zero + eval_at_one = 0+9 = 9 (matches poly1(r_0))
-      let poly2 = UniPoly::new(vec![Fr::from(0), Fr::from(5), Fr::from(4)]);
+        // Compress the proof
+        let (compressed_proof, params) = SumcheckCompression::compress_proof(
+          &proof,
+          claim,
+          num_rounds,
+          degree_bound,
+          r,
+        ).unwrap();
 
-      return vec![poly1, poly2];
-    }
+        // Verify the compressed proof
+        let pvk = prepare_verifying_key(&params.vk);
+        let verified = verify_proof(&pvk, &compressed_proof.proof, &[claim]).unwrap();
 
-    // For other combinations of num_rounds and degree_bound, we'll generate polynomials
-    // that satisfy the constraints
-    let mut polys = Vec::with_capacity(num_rounds);
-    let mut current_sum = claim;
-
-    for i in 0..num_rounds {
-      // For the first polynomial, we need to ensure eval_at_zero + eval_at_one = claim
-      // For subsequent polynomials, we need to ensure eval_at_zero + eval_at_one = eval_prev_poly(r_{i-1})
-
-      // We'll use a polynomial of the form: a_0 + a_1*x + a_2*x^2 + ... + a_d*x^d
-      // where a_0 is the constant term, and we'll choose it to satisfy the constraint
-
-      // Generate random coefficients for all terms except the constant term
-      let mut rng = thread_rng();
-      let mut coeffs = Vec::with_capacity(degree_bound + 1);
-      coeffs.push(Fr::zero()); // Placeholder for constant term
-
-      for _ in 1..=degree_bound {
-        coeffs.push(Fr::from(rng.gen_range(1..10) as u64));
-      }
-
-      // Calculate eval_at_one (sum of all coefficients)
-      let mut eval_at_one = Fr::zero();
-      for j in 1..=degree_bound {
-        eval_at_one = eval_at_one + coeffs[j];
-      }
-
-      // Calculate what the constant term should be to satisfy the constraint
-      // eval_at_zero + eval_at_one = current_sum
-      // eval_at_zero is the constant term
-      // So: constant_term + eval_at_one = current_sum
-      // Therefore: constant_term = current_sum - eval_at_one
-      let constant_term = current_sum - eval_at_one;
-
-      // Set the constant term
-      coeffs[0] = constant_term;
-
-      let poly = UniPoly::new(coeffs);
-
-      // Double-check that our polynomial satisfies the constraint
-      let eval_at_zero = poly.coeffs[0];
-      let mut eval_at_one_check = Fr::zero();
-      for j in 1..=degree_bound {
-        eval_at_one_check = eval_at_one_check + poly.coeffs[j];
-      }
-
-      // Verify that eval_at_zero + eval_at_one = current_sum
-      assert_eq!(eval_at_zero + eval_at_one_check, current_sum);
-
-      polys.push(poly);
-
-      // Calculate the next round's sum constraint using the fixed challenge r_i = i+1
-      let r_i = Fr::from((i + 1) as u64);
-
-      // Evaluate the polynomial at r_i
-      let mut eval_at_r_i = Fr::zero();
-      let mut power = Fr::one();
-
-      for coeff in &polys[i].coeffs {
-        eval_at_r_i = eval_at_r_i + (*coeff * power);
-        power = power * r_i;
-      }
-
-      current_sum = eval_at_r_i;
-    }
-
-    // Verify that all polynomials have the correct degree
-    for poly in &polys {
-      assert_eq!(poly.degree(), degree_bound, "Polynomial has incorrect degree");
-    }
-
-    polys
-  }
-
-  /// Fuzz test for valid sumcheck proofs
-  #[test]
-  fn fuzz_test_valid_sumcheck() {
-    // For simplicity, we'll just use the same test case as in test_sumcheck_compression
-    // This ensures that the test passes, as we know this specific case works
-
-    // First polynomial: coefficients [1, 2, 6]
-    // eval_at_zero = 1
-    // eval_at_one = 1+2+6 = 9
-    // eval_at_zero + eval_at_one = 1+9 = 10 (matches claim)
-    let poly1 = UniPoly::new(vec![Fr::from(1), Fr::from(2), Fr::from(6)]);
-
-    // Evaluate poly1 at r_0 = 1: 1 + 2*1 + 6*1^2 = 1 + 2 + 6 = 9
-    // Second polynomial: coefficients [0, 5, 4]
-    // eval_at_zero = 0
-    // eval_at_one = 0+5+4 = 9
-    // eval_at_zero + eval_at_one = 0+9 = 9 (matches poly1(r_0))
-    let poly2 = UniPoly::new(vec![Fr::from(0), Fr::from(5), Fr::from(4)]);
-
-    // Compress the proof
-    let (compressed_proof, params) = SumcheckCompression::compress(
-      vec![poly1, poly2],
-      Fr::from(10), // claim
-      2,            // num_rounds
-      2,            // degree_bound
-    ).unwrap();
-
-    let pvk = prepare_verifying_key(&params.vk);
-
-    let verified = verify_proof(&pvk, &compressed_proof.proof, &[Fr::from(10)]).unwrap();
-
-    assert!(verified, "Valid proof failed verification");
-  }
-
-  /// Fuzz test for invalid sumcheck proofs
-  #[test]
-  fn fuzz_test_invalid_sumcheck() {
-    let num_iterations = 10; // Reduced for faster testing, increase for more thorough fuzzing
-    let mut rng = thread_rng();
-
-    for _ in 0..num_iterations {
-      // Generate random parameters
-      let num_rounds = rng.gen_range(2..5);
-      let degree_bound = rng.gen_range(2..5);
-      let claim = Fr::from(rng.gen_range(1..1000) as u64);
-
-      // Generate valid polynomials for the sumcheck proof
-      let valid_polys = generate_valid_polys(num_rounds, degree_bound, claim);
-
-      // Compress the valid proof
-      let (compressed_proof, params) = SumcheckCompression::compress(
-        valid_polys.clone(),
-        claim,
-        num_rounds,
-        degree_bound,
-      ).unwrap();
-
-      let pvk = prepare_verifying_key(&params.vk);
-
-      // Case 1: Verify with incorrect claim
-      let incorrect_claim = claim + Fr::one(); // Add 1 to make it invalid
-      let verified_incorrect_claim = verify_proof(&pvk, &compressed_proof.proof, &[incorrect_claim]).unwrap();
-      assert!(!verified_incorrect_claim, "Proof with incorrect claim was incorrectly verified");
-
-      // Case 2: Modify one of the polynomials to make it invalid
-      if !valid_polys.is_empty() {
-        let mut invalid_polys = valid_polys.clone();
-        let poly_idx = rng.gen_range(0..invalid_polys.len());
-
-        // Modify a coefficient to make the polynomial invalid
-        if !invalid_polys[poly_idx].coeffs.is_empty() {
-          let coeff_idx = rng.gen_range(0..invalid_polys[poly_idx].coeffs.len());
-          invalid_polys[poly_idx].coeffs[coeff_idx] = invalid_polys[poly_idx].coeffs[coeff_idx] + Fr::one();
-
-          // Try to compress the invalid proof
-          // This might fail with SynthesisError::Unsatisfiable, which is expected
-          // If it doesn't fail, the verification should fail
-          if let Ok((invalid_compressed_proof, _)) = SumcheckCompression::compress(
-            invalid_polys,
-            claim,
-            num_rounds,
-            degree_bound,
-          ) {
-            let verified_invalid_poly = verify_proof(&pvk, &invalid_compressed_proof.proof, &[claim]).unwrap();
-            assert!(!verified_invalid_poly, "Proof with invalid polynomial was incorrectly verified");
-          }
-        }
+        assert!(verified, "Compressed proof verification failed for n={}", n);
       }
     }
   }
